+++ CHANGELOG.md
@@ -17,6 +17,10 @@ Exceptions are acceptable depending on the circumstances (critical bug fixes tha
 
 ## [Unreleased]
 
+### Removed
+
+- removed `IsUnknown` from inside the `ValidateConfig` method to avoid issues in to apply without previous state
+
 ## [1.2.0] - 2024-12-09
 
 ### Added